+++ cmd/neofs-rest-gw/config.go
@@ -17,8 +17,6 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/wallet"
 	"github.com/nspcc-dev/neofs-rest-gw/handlers"
 	"github.com/nspcc-dev/neofs-rest-gw/metrics"
-	"github.com/nspcc-dev/neofs-sdk-go/client"
-	"github.com/nspcc-dev/neofs-sdk-go/netmap"
 	"github.com/nspcc-dev/neofs-sdk-go/pool"
 	"github.com/nspcc-dev/neofs-sdk-go/user"
 	"github.com/spf13/pflag"
@@ -37,9 +35,6 @@ const (
 
 	defaultPoolErrorThreshold   uint32 = 100
 	defaultPoolDefaultTimestamp bool   = false
-
-	// default [handlers.PrmAPI.MaxPayloadBufferSize] value.
-	defaultMaxObjectPayloadBufferSize = 4 << 20
 
 	// Pool config.
 	cmdNodeDialTimeout      = "node-dial-timeout"
@@ -583,11 +578,6 @@ func newNeofsAPI(ctx context.Context, logger *zap.Logger, v *viper.Viper) (*hand
 		return nil, err
 	}
 
-	ni, err := getNetworkInfo(ctx, logger, p, len(peers))
-	if err != nil {
-		return nil, err
-	}
-
 	var apiPrm handlers.PrmAPI
 	apiPrm.Pool = p
 	apiPrm.Key = key
@@ -605,36 +595,9 @@ func newNeofsAPI(ctx context.Context, logger *zap.Logger, v *viper.Viper) (*hand
 	}
 
 	apiPrm.ServiceShutdownTimeout = defaultShutdownTimeout
-	if apiPrm.MaxPayloadBufferSize = ni.MaxObjectSize(); apiPrm.MaxPayloadBufferSize == 0 {
-		// default to some heuristic value that in practice should not be needed at all
-		apiPrm.MaxPayloadBufferSize = defaultMaxObjectPayloadBufferSize
-		logger.Debug("NeoFS max object size setting is zero, using default limit for payload buffer size",
-			zap.Uint64("value", apiPrm.MaxPayloadBufferSize))
-	}
-
 	apiPrm.DefaultTimestamp = v.GetBool(cfgPoolDefaultTimestamp)
 
 	return handlers.NewAPI(&apiPrm)
-}
-
-func getNetworkInfo(ctx context.Context, logger *zap.Logger, p *pool.Pool, peerCount int) (netmap.NetworkInfo, error) {
-	var retries = max(peerCount, 1)
-
-	for range retries {
-		ni, err := p.NetworkInfo(ctx, client.PrmNetworkInfo{})
-		if err == nil {
-			return ni, nil
-		}
-		logger.Warn("failed to get network info", zap.Error(err))
-
-		select {
-		case <-ctx.Done():
-			return ni, fmt.Errorf("networkInfo: interrupted: %w", err)
-		default:
-		}
-	}
-
-	return netmap.NetworkInfo{}, fmt.Errorf("get networkInfo failed after %d retries", retries)
 }
 
 func fetchPeers(l *zap.Logger, v *viper.Viper) []pool.NodeParam {